move 3 auth methods to base class, add api login
test for api setup user
